+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - IPv6 configuration in the example FRR config.
+- Full solution diagram.
 
 ## [1.3.0] - 2025-05-08